backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
python-multipart>=0.0.18
boto3>=1.35.0
httpx>=0.28.0
redis>=5.0.0
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-exporter-otlp-proto-grpc>=1.25.0
PyJWT>=2.9.0
bcrypt>=4.2.0
pytest>=8.3.0
pytest-asyncio>=0.25.0
